+++ csatar-plugins/csatar/models/AccidentLogRecord.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Csatar\Csatar\Models;
+
+use Model;
+
+class AccidentLogRecord extends Model
+{
+
+}